chore: 설정 변경, 시퀀스 스키마 변경

+++ src/main/resources/schema.sql
@@ -1,3 +1,3 @@
 /* 환자등록번호 채번용 시퀀스 */
 drop sequence if exists patient_registration_number_sequence;
-create sequence patient_registration_number_sequence as bigint minvalue 1 maxvalue 9999999999999 increment by 1;
+create sequence patient_registration_number_sequence minvalue 1 maxvalue 9999999999999 increment by 1;
